+++ packages/components/CHANGELOG.md
@@ -20,6 +20,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   settings ([#2719](https://github.com/porsche-design-system/porsche-design-system/pull/2719))
 - `Pin Code` ([#2691](https://github.com/porsche-design-system/porsche-design-system/pull/2691))
 
+#### Changed
+
+- Usage of `getInitialStyles()` partial is required and validated with an exception
+  ([#2749](https://github.com/porsche-design-system/porsche-design-system/pull/2749))
+
 ### [3.7.0-rc.0] - 2023-09-05
 
 #### Added